Use constants, NOT strings.
TIL:
If you invoke instance doubles in RSpec with a string argument, instead
of a constant, if the constant isn't loaded yet, rspec will let you
'slide' - i.e. it won't enforce your interfaces....which is kind of the
WHOLE POINT of instance double vs double.  Because of that, use the
constant as the first arguement, not a string where I can.