Fix issues with service areas.
It seems some carries issued service areas at both the state and
county/zip level - and then gave them the same service area identifier.
This corrects the creation of service areas from the original data to be
biased in favor of the state-wide areas when the carrier for some reason
provided both.