Correct criteria for preventing transmission of enrollment events. (#2854)
The current check examines if we offer the plan attached to the enrollment.
However, it uses the **current** address of the covered individuals. Not
the address which was present at the purchase of the enrollment.
This means that when you move, resulting in a termination, the termination
event will no longer flow, because all of a sudden, the plan isn't offered -
which is correct.  What we need to check for instead is an **actual**
$0 premium.

Co-authored-by: vishal kalletla <vishuk1201@gmail.com>